Fall risk API works locally, and scales parameter values for age, berg, and gait speed
